Template Railtype
Okay, I was cranky yesterday; If you need  me to reconsider, all I request is that you understand it wasn't my own fault, and that I recognize you want this sort of drama to not be in your community. Hence, why a "declaration of war" is in my commit comments and not the community Discord.